create categoriy button instead of dropdown
